refactor(chat): move getConversations into lib/chat/
Aligns with the repo convention (CLAUDE.md): "File Organization —
Domain-specific directories (e.g. /lib/fal/ not /lib/utils/fal.ts)".

- git mv lib/getConversations.tsx       → lib/chat/getConversations.tsx
- git mv lib/__tests__/getConversations.test.ts → lib/chat/__tests__/getConversations.test.ts
- Updated the one importer (hooks/useConversations.tsx) and the
  test's relative consts import.

No behavior change; tests + imports verified.
